test-repo/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.17
langchain-openai==0.1.8
langchain-community==0.0.38
langchain-core==0.1.52
chromadb==0.4.21
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
pandas==2.1.4
python-multipart==0.0.6
python-dateutil==2.8.2 